refactor: split utilities into functional files
