+++ src/utils/depot-tools-utils.js
@@ -1,0 +1,90 @@
+const fs = require('fs');
+const path = require('path');
+const os = require('os');
+const childProcess = require('child_process');
+const pathKey = require('path-key');
+
+const { color } = require('./e-utils');
+
+const defaultDepotPath = path.resolve(__dirname, '..', 'third_party', 'depot_tools');
+const DEPOT_TOOLS_DIR = process.env.DEPOT_TOOLS_DIR || defaultDepotPath;
+
+function updateDepotTools() {
+  const depot_dir = DEPOT_TOOLS_DIR;
+  console.log(`Updating ${color.path(depot_dir)}`);
+  if (os.platform() === 'win32') {
+    depotExecFileSync({}, 'cmd.exe', ['/c', path.resolve(depot_dir, 'update_depot_tools.bat')]);
+  } else {
+    depotExecFileSync({}, path.resolve(depot_dir, 'update_depot_tools'));
+  }
+}
+
+function ensureDepotTools() {
+  const depot_dir = DEPOT_TOOLS_DIR;
+
+  // if it doesn't exist, create it
+  if (!fs.existsSync(depot_dir)) {
+    console.log(`Cloning ${color.cmd('depot_tools')} into ${color.path(depot_dir)}`);
+    const url = 'https://chromium.googlesource.com/chromium/tools/depot_tools.git';
+    childProcess.execFileSync('git', ['clone', '-q', url, depot_dir], { stdio: 'inherit' });
+    updateDepotTools();
+  }
+
+  // if it's been awhile, update it
+  const now = new Date();
+  const msec_per_day = 86400000;
+  const days_before_pull = 14;
+  const days_untouched = (now.getTime() - fs.statSync(depot_dir).mtimeMs) / msec_per_day;
+  if (days_untouched >= days_before_pull) {
+    updateDepotTools();
+    fs.utimesSync(depot_dir, now, now);
+  }
+}
+
+function depotOpts(config, opts = {}) {
+  // some defaults
+  opts = {
+    encoding: 'utf8',
+    stdio: 'inherit',
+    ...opts,
+  };
+
+  opts.env = {
+    // set these defaults that can be overridden via process.env
+    PYTHONDONTWRITEBYTECODE: '1', // depot needs it
+    DEPOT_TOOLS_METRICS: '0', // disable depot metrics
+    ...process.env,
+    ...config.env,
+    ...opts.env,
+  };
+
+  // put depot tools at the front of the path
+  const key = pathKey();
+  opts.env[key] = [DEPOT_TOOLS_DIR, process.env[key]].join(path.delimiter);
+
+  return opts;
+}
+
+function depotExecSync(config, cmd, opts_in) {
+  const opts = depotOpts(config, opts_in);
+  console.log(color.childExec(cmd, null, opts));
+  childProcess.execSync(cmd, opts);
+}
+
+function depotExecFileSync(config, exec, args, opts_in) {
+  if (exec === 'python' && !path.isAbsolute(args[0])) {
+    args[0] = path.resolve(DEPOT_TOOLS_DIR, args[0]);
+  }
+  const opts = depotOpts(config, opts_in);
+  console.log(color.childExec(exec, args, opts));
+  childProcess.execFileSync(exec, args, opts);
+}
+
+module.exports = {
+  depot: {
+    path: DEPOT_TOOLS_DIR,
+    ensure: ensureDepotTools,
+    execFileSync: depotExecFileSync,
+    execSync: depotExecSync,
+  }
+}
