+++ src/utils/macos-sdks-utils.js
@@ -1,0 +1,23 @@
+const childProcess = require('child_process');
+const path = require('path');
+
+const { color } = require('./e-utils');
+
+const macOSSDKsPath = path.resolve(__dirname, '..', 'third_party', 'macOS_SDKs');
+
+function ensureMacOSSDKs() {
+  if (!fs.existsSync(macOSSDKsPath)) {
+    console.log(`Cloning ${color.cmd('MacOSX-SDKs')} into ${color.path(macOSSDKsPath)}`);
+    const url = 'https://github.com/phracker/MacOSX-SDKs.git';
+    childProcess.execFileSync('git', ['clone', '-q', url, macOSSDKsPath], { stdio: 'inherit' });
+  } else {
+    childProcess.execFileSync('git', ['pull', '-q'], { cwd: macOSSDKsPath, stdio: 'inherit' });
+  }
+}
+
+module.exports = {
+  macOSSDKs: {
+    path: macOSSDKsPath,
+    ensure: ensureMacOSSDKs,
+  }
+}
