+++ tests/e-show-spec.js
@@ -1,4 +1,3 @@
-const fs = require('fs');
 const path = require('path');
 const createSandbox = require('./sandbox');
