+++ tests/sandbox.js
@@ -1,6 +1,5 @@
 const childProcess = require('child_process');
 const fs = require('fs');
-const os = require('os');
 const path = require('path');
 
 // for `rm -rf`'ing the sandbox tmpdir
